Move contract_common to contracts
