+++ gnosis/eth/contracts/contract_common.py
